@@ -2,6 +2,7 @@ package Model;
 
 import java.io.File;
 import java.text.SimpleDateFormat;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
@@ -9,25 +10,30 @@ import java.util.HashMap;
 public class CodedOff extends SaveAble {
     private static ArrayList<CodedOff> allOfCodes = new ArrayList<>();
     private String offBarcode;
-    private SimpleDateFormat startTime;
-    private SimpleDateFormat endTime;
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
     private double offAmount;
     private int usageTime;
+    private int percent;
     private HashMap<Account, Integer> numberOfUsageForEachAccount;
     private ArrayList<Account> discountIsForTheseAccounts;
     private ArrayList<Customer> containingCustomers;
-    private static ArrayList<String> offBarcodes = new ArrayList<>();
+   // private static ArrayList<String> offBarcodes = new ArrayList<>();
 
-    public CodedOff(String offBarcode, SimpleDateFormat startTime, SimpleDateFormat endTime, double offAmount, int usageTime, ArrayList<Customer> containingCustomers) {
+    public CodedOff(String offBarcode, LocalDateTime startTime, LocalDateTime endTime, double offAmount, int percent, int usageTime, ArrayList<Customer> containingCustomers) {
         this.offBarcode = offBarcode;
         this.startTime = startTime;
         this.endTime = endTime;
         this.offAmount = offAmount;
+        this.percent = percent;
         this.usageTime = usageTime;
         this.containingCustomers = containingCustomers;
         this.numberOfUsageForEachAccount = new HashMap<Account, Integer>();
         this.discountIsForTheseAccounts = new ArrayList<Account>();
+        System.err.println(allOfCodes);
         allOfCodes.add(this);
+        System.err.println(allOfCodes);
+        SaveAndLoad.getSaveAndLoad().writeJSON(allOfCodes, ArrayList.class, "offCodes");
     }
 
     public static void removeOffCode(CodedOff offCode) {
@@ -44,9 +50,9 @@ public class CodedOff extends SaveAble {
         allOfCodes.remove(offCode);
     }
 
-    public static ArrayList<String> getOffBarcodes() {
-        return offBarcodes;
-    }
+    ///public static ArrayList<String> getOffBarcodes() {
+       // return offBarcodes;
+    //}
 
     public String getOffBarcode() {
         return offBarcode;
